Fixing backgrounds.
