+++ api_v2/serializers/__init__.py
@@ -6,3 +6,6 @@ from .item import ArmorSerializer
 from .item import WeaponSerializer
 from .item import ItemSerializer
 from .item import ItemSetSerializer
+
+from .background import BackgroundBenefitSerializer
+from .background import BackgroundSerializer
