+++ api_v2/serializers/serializers.py
@@ -289,18 +289,3 @@ class CreatureSerializer(GameContentSerializer):
         return result
 
 
-class BackgroundBenefitSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.BackgroundBenefit
-        fields = ['name','desc','type']
-
-
-class BackgroundSerializer(GameContentSerializer):
-    key = serializers.ReadOnlyField()
-    benefits = BackgroundBenefitSerializer(
-        many=True
-    )
-
-    class Meta:
-        model = models.Background
-        fields = '__all__'
